Bug fix: issue when adding features
There must have been some sort of behavior change with the ordering of layers in the map. I don't think this qualifies as a bug in dependencies, since the code was making an invalid assumption that happened to work previously.

Previously, the code to add a feature would add to the first feature layer found, and throw an exception if that didn't work.

Now, the add method attempts to add to the first feature layer with point geometry. If there is an error, it keeps trying the rest of the point feature layers. Operation completes if feature can be added to any point layer. If feature cannot be added to any layer, an aggregate exception will be raised with any exceptions from adding to individual layers.

This isn't the most comprehensive solution, as it assumes there will always be at least one point feature layer. I think it makes sense to go with this temporary solution, since we will be updating to add multiple geometry support very soon, and that will require a complete rewrite of this method.